Adding back availability to fold in playlists
Re-enabled playlist request for playback, but using a qparam:

`playlists=1`

Append that param to the page (e.g., `playback.jsp?playlists=1&view=hls`) and
playlists will be requested (as well as try to force HLS as tech playback).

Be sure to look at console, because if you force HLS playback and no playlists are returned, playback of FLVs will not work in the front-end; take off `view=hls` to test all failover.